requirements.txt
ffmpeg-python
elasticsearch
urllib3
python-dotenv
requests
tabulate
Retry
psycopg2-binary
aiogram
asyncpg
pydantic-settings
pydantic